Moved Schema replace vars to its own integration.
