+++ src/generators/schema-generator.php
@@ -133,10 +133,12 @@ class Schema_Generator implements Generator_Interface {
 				$graph      = \apply_filters( 'wpseo_schema_block_' . $block_type, $graph, $block, $context );
 
 				if ( isset( $block['attrs']['yoast-schema'] ) ) {
-					$graph[] = $this->replace_vars( $block['attrs']['yoast-schema'], $context->presentation );
+					$graph[] = $block['attrs']['yoast-schema'];
 				}
 			}
 		}
+
+		$graph = $this->replace_vars( $graph, $context->presentation );
 
 		return [
 			'@context' => 'https://schema.org',
